backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.10.0
python-multipart==0.0.22
pandas==2.2.0
slowapi==0.1.9
charset-normalizer
redis